requirements.txt
tornado==5.1.1
pyeva==3.0.1
sqlalchemy==1.2.12
sqlalchemy-utils==0.33.5
psycopg2-binary==2.7.5
